chore: refactor code structure for improved readability and maintainability
