+++ packages/frontend/app/(tabs)/explore.web.tsx
